refactor(app): split the poll loop from its reads
app.go held four unrelated things at once: the loop, the backoff that
decides how loudly a run of failures is logged, the per-pane reads, and
the assembly of tabs from a snapshot. Reading any one of them meant
scrolling past the other three, and the rename sat four levels deep
inside the loop that decided on it.

The loop keeps app.go. failureLog moves to failures.go, where nothing
ties it to App. The reads and the tab assembly move to reads.go. The
rename itself becomes App.rename, leaving the loop a list of reasons to
skip a tab and one line that acts.

Code and comments move verbatim; behaviour is unchanged.
